requirements.txt
Django==1.5
PIL==1.1.7
boto==2.8.0
distribute==0.6.34
django-piston==0.2.3
django-storages==1.1.6
html5lib==0.95
isodate==0.4.9
psycopg2==2.4.6
pyparsing==1.5.7
wsgiref==0.1.2
git+https://github.com/RDFLib/rdflib.git
git+https://github.com/RDFLib/rdflib-postgresql.git
git+https://github.com/RDFLib/rdflib-sparql.git
